# openapi/openapi.yaml
openapi: 3.0.3
info:
  description: Th1nk
  title: Th1nk
  version: 1.28.0
servers:
- description: Local server
  url: "http://localhost:{port}/api/v1"
  variables:
    port:
      default: "8080"
- description: Dev server
  url: "http://pg-doener-dev.virt.uni-oldenburg.de:{port}/api/v1"
  variables:
    port:
      default: "8080"
- description: Prod server
  url: "http://pg-doener-prod.virt.uni-oldenburg.de:{port}/api/v1"
  variables:
    port:
      default: "8080"
paths:
  /table-structures:
    get:
      operationId: getTableStructures
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/tableStructureSummary'
                type: array
          description: OK
        "401":
          description: Unauthorized
      security:
      - oAuth2Auth:
        - read:tablestructure
      summary: Get all table structures
    post:
      operationId: createTableStructure
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tableStructure'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                format: int64
                minimum: 1
                type: integer
          description: Created
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      security:
      - oAuth2Auth:
        - write:tablestructure
      summary: Create a new table structure
  /table-structures/{id}:
    delete:
      operationId: deleteTableStructure
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int64
          minimum: 1
          type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
      security:
      - oAuth2Auth:
        - write:tablestructure
      summary: Delete table structure by id
    get:
      operationId: getTableStructure
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int64
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tableStructure'
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
      security:
      - oAuth2Auth:
        - read:tablestructure
      summary: Get a table structure by id
    put:
      operationId: updateTableStructure
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int64
          minimum: 1
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tableStructure'
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
      security:
      - oAuth2Auth:
        - write:tablestructure
      summary: Update table structure by id
  /table-structures/generate:
    post:
      operationId: generateTableStructure
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/generateTableStructure_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tableStructureGenerationResponse'
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      security:
      - oAuth2Auth:
        - read:tablestructure
      summary: Generate a tableStructure and return the result as a json
  /converter/{tableStructureId}:
    post:
      operationId: convertTable
      parameters:
      - in: path
        name: tableStructureId
        required: true
        schema:
          format: int64
          minimum: 1
          type: integer
      - in: query
        name: mode
        required: false
        schema:
          default: CREATE
          enum:
          - CREATE
          - EXTEND
          - REPLACE
          - RESTRUCTURE
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/convertTable_request'
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
      security:
      - oAuth2Auth:
        - write:converter
      summary: Convert a table and save it to the database
  /converter/preview:
    post:
      operationId: previewConvertTable
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          minimum: 1
          type: integer
      - in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/previewConvertTable_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  items:
                    type: string
                  type: array
                type: array
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      security:
      - oAuth2Auth:
        - read:converter
      summary: Convert a table and return a preview of the result
  /converter/file:
    post:
      operationId: fileConvertTable
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/previewConvertTable_request'
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      security:
      - oAuth2Auth:
        - read:converter
      summary: Convert a table and return the result as a file
...
components:
  schemas:
    tableStructure:
      description: |
        This is a table structure.
        It contains all structures and their information needed to convert a table.
        The properties endRow and endColumn can be used to cut off rows and columns at the end of the table.
      example:
        id: 1
        name: Table structure 1
        structures:
        - name: Fill Header
          description: Fills empty cells in the header
          converterType: FILL_EMPTY_ROW
          rowIndex:
          - 0
          - 1
        - name: Resolve Grouped Header
          description: Resolves the grouped header
          converterType: REMOVE_GROUPED_HEADER
          rowIndex:
          - 0
          - 1
          columnIndex:
          - 0
          startRow: 3
          startColumn: 0
      properties:
        id:
          description: The ID of the table structure.
          format: int64
          minimum: 1
          type: integer
        name:
          description: |
            The name of the table structure.
            This is an arbitrary string.
          example: Table structure 1
          minLength: 1
          type: string
        structures:
          description: The list of structures in the table structure.
          items:
            $ref: '#/components/schemas/structure'
          minItems: 0
          type: array
        endRow:
          description: |
            The last zero-based row index of the table.
            All rows after this index are removed.
            If this property is not set, all rows are kept.
          minimum: 1
          type: integer
        endColumn:
          description: |
            The last zero-based column index of the table.
            All columns after this index are removed.
            If this property is not set, all columns are kept.
          minimum: 1
          type: integer
      required:
      - delimiter
      - name
      - structures
      type: object
    structure:
      description: |
        This is the base for all converter structures.
        The converterType is used as a discriminator to determine the specific structure.
      discriminator:
        mapping:
          REMOVE_GROUPED_HEADER: '#/components/schemas/removeGroupedHeaderStructure'
          FILL_EMPTY_ROW: '#/components/schemas/fillEmptyRowStructure'
          FILL_EMPTY_COLUMN: '#/components/schemas/fillEmptyColumnStructure'
          REMOVE_COLUMN_BY_INDEX: '#/components/schemas/removeColumnByIndexStructure'
          REMOVE_ROW_BY_INDEX: '#/components/schemas/removeRowByIndexStructure'
          ADD_HEADER_NAME: '#/components/schemas/addHeaderNameStructure'
          REMOVE_HEADER: '#/components/schemas/removeHeaderStructure'
          REMOVE_FOOTER: '#/components/schemas/removeFooterStructure'
          REPLACE_ENTRIES: '#/components/schemas/replaceEntriesStructure'
          SPLIT_CELL: '#/components/schemas/splitCellStructure'
          REMOVE_INVALID_ROWS: '#/components/schemas/removeInvalidRowsStructure'
          REMOVE_TRAILING_COLUMN: '#/components/schemas/removeTrailingColumnStructure'
          REMOVE_LEADING_COLUMN: '#/components/schemas/removeLeadingColumnStructure'
          MERGE_COLUMNS: '#/components/schemas/mergeColumnsStructure'
          TRANSPOSE_MATRIX: '#/components/schemas/transposeMatrixStructure'
          PIVOT_MATRIX: '#/components/schemas/pivotMatrixStructure'
          REMOVE_KEYWORD: '#/components/schemas/removeKeywordsStructure'
        propertyName: converterType
      example:
        name: Resolve Grouped Header
        description: Resolves the grouped header
        converterType: REMOVE_GROUPED_HEADER
        rowIndex:
        - 0
        - 1
        columnIndex:
        - 0
        startRow: 3
        startColumn: 0
      oneOf:
      - $ref: '#/components/schemas/removeRowByIndexStructure'
      - $ref: '#/components/schemas/removeColumnByIndexStructure'
      - $ref: '#/components/schemas/removeGroupedHeaderStructure'
      - $ref: '#/components/schemas/fillEmptyRowStructure'
      - $ref: '#/components/schemas/fillEmptyColumnStructure'
      - $ref: '#/components/schemas/addHeaderNameStructure'
      - $ref: '#/components/schemas/removeHeaderStructure'
      - $ref: '#/components/schemas/removeFooterStructure'
      - $ref: '#/components/schemas/replaceEntriesStructure'
      - $ref: '#/components/schemas/splitCellStructure'
      - $ref: '#/components/schemas/removeInvalidRowsStructure'
      - $ref: '#/components/schemas/removeTrailingColumnStructure'
      - $ref: '#/components/schemas/removeLeadingColumnStructure'
      - $ref: '#/components/schemas/mergeColumnsStructure'
      - $ref: '#/components/schemas/transposeMatrixStructure'
      - $ref: '#/components/schemas/pivotMatrixStructure'
      - $ref: '#/components/schemas/removeKeywordsStructure'
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        name:
          description: |
            A name for the structure.
            This can be used to identify the structure in the UI.
            It can be any string.
          minLength: 1
          type: string
        description:
          description: |
            A description for the structure.
            This can be used to provide more information about the structure in the UI.
            It can be any string.
          minLength: 1
          type: string
      required:
      - converterType
      type: object
    tableStructureSummary:
      description: |
        This is a summary of a table structure.
        It contains am ID, name, and an array of structure summaries.
        However, it does not contain the actual structure with its fields.
      example:
        name: name
        structures:
        - converterType: REMOVE_GROUPED_HEADER
        - converterType: REMOVE_GROUPED_HEADER
        id: 1
      properties:
        id:
          description: The ID of the table structure.
          format: int64
          minimum: 1
          type: integer
        name:
          description: |
            The name of the table structure.
            This is an arbitrary string.
          minLength: 1
          type: string
        structures:
          description: The list of structures in the table structure as summaries.
          items:
            $ref: '#/components/schemas/structureSummary'
          minItems: 0
          type: array
      required:
      - id
      - name
      - structures
      type: object
    structureSummary:
      description: |
        This is a summary of a structure.
        It only contains the converterType to identify the use of the structure.
      example:
        converterType: REMOVE_GROUPED_HEADER
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
      required:
      - converterType
      type: object
    converterType:
      description: This enum is used as a discriminator for structures.
      enum:
      - REMOVE_GROUPED_HEADER
      - FILL_EMPTY_ROW
      - FILL_EMPTY_COLUMN
      - REMOVE_COLUMN_BY_INDEX
      - REMOVE_ROW_BY_INDEX
      - ADD_HEADER_NAME
      - REMOVE_FOOTER
      - REMOVE_HEADER
      - REPLACE_ENTRIES
      - SPLIT_CELL
      - REMOVE_INVALID_ROWS
      - REMOVE_TRAILING_COLUMN
      - REMOVE_LEADING_COLUMN
      - MERGE_COLUMNS
      - TRANSPOSE_MATRIX
      - PIVOT_MATRIX
      - REMOVE_KEYWORD
      type: string
      x-field-extra-annotation: '@de.uol.pgdoener.th1.api.validation.ValidConverterType'
    removeRowByIndexStructure:
      description: |
        This converter removes the specified rows from the table.
        If the specified row index is out of bounds, an error will be thrown.

        The converterType MUST be REMOVE_ROW_BY_INDEX.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        rowIndex:
          description: The indices of the rows to remove
          items:
            minimum: 0
            type: integer
          minItems: 1
          type: array
      required:
      - converterType
      - rowIndex
      type: object
    removeColumnByIndexStructure:
      description: |
        This converter removes the specified columns from the table.
        If the specified column index is out of bounds, an error will be thrown.

        The converterType MUST be REMOVE_COLUMN_BY_INDEX.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        columnIndex:
          description: The indices of the columns to remove
          items:
            minimum: 0
            type: integer
          minItems: 1
          type: array
      required:
      - columnIndex
      - converterType
      type: object
    removeGroupedHeaderStructure:
      description: |
        This converter restructures the table to remove the grouped header.

        The rowIndex list contains the row indices of the grouped header rows.
        This can be used to describe grouped headers at the top of the table.
        Rows above the grouped header are not allowed, thus the first index in the rowIndex list has to be 0.

        The columnIndex list contains the column indices of the grouped header columns.
        Column indices can be used to describe grouped headers at the left of the table.
        All columns to the left of the lowest column index are removed.

        The properties startRow and startColumn can be used to specify the topmost and leftmost cell data points below and to the left of the grouped header.
        The startRow and startColumn properties are optional and default to below last rowIndex and to the right of the last columnIndex.

        The converterType MUST be REMOVE_GROUPED_HEADER.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        rowIndex:
          description: The indices of the rows that are part of the grouped header
            at the top of the table.
          items:
            minimum: 0
            type: integer
          minItems: 1
          type: array
        columnIndex:
          description: The indices of the columns that are part of the grouped header
            at the left of the table.
          items:
            minimum: 0
            type: integer
          minItems: 1
          type: array
        startRow:
          description: The index of the first data row below the grouped header.
          minimum: 0
          type: integer
        startColumn:
          description: The index of the first data column to the right of the grouped
            header.
          minimum: 0
          type: integer
      required:
      - columnIndex
      - converterType
      - rowIndex
      type: object
    addHeaderNameStructure:
      description: |
        This converter replaces the first row of the table with the specified header names.
        If the number of provided header names is less than the number of columns in the table, the remaining columns will not be changed.
        If the number of provided header names is greater than the number of columns in the table, an error will be thrown.

        Two modes are supported:
        1. REPLACE_FIRST_ROW: Replaces the row at a given index with the specified headers.
        2. INSERT_AT_TOP: Shifts the table down by one row and inserts the headers as the new first row.
        The converterType MUST be ADD_HEADER_NAME.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        headerNames:
          description: The header names
          example:
          - Name
          - Date of Birth
          items:
            minLength: 1
            type: string
          minItems: 1
          type: array
        headerPlacementType:
          $ref: '#/components/schemas/header_placement_type'
      required:
      - converterType
      - headerNames
      - headerPlacementType
      type: object
...
    fillEmptyRowStructure:
      description: |
        This converter fills empty cells in the specified rows.
        It iterates through each specified row and fills empty cells with the value from the last non-empty cell.
        If there are no non-empty cells in the row an error will be thrown.
        This converter expects a non-empty cell at the beginning of each specified row.

        The converterType MUST be FILL_EMPTY_ROW.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        rowIndex:
          description: The indices of the rows to fill
          items:
            minimum: 0
            type: integer
          minItems: 1
          type: array
      required:
      - converterType
      - rowIndex
      type: object
    fillEmptyColumnStructure:
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        columnIndex:
          items:
            minimum: 0
            type: integer
          minItems: 0
          type: array
      required:
      - columnIndex
      - converterType
      type: object
    removeHeaderStructure:
      description: |
        This converter removes the header from the table.
        The header is defined as the first rows of the table which only contain a maximum of threshold invalid values.
        The invalid values are empty cells by default but can be extended using the blockList properties.

        The converterType MUST be REMOVE_HEADER.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        threshold:
          description: The maximum number of invalid values in a row to be considered
            a header row.
          minimum: 0
          type: integer
        blockList:
          description: The list of values to be considered invalid.
          items:
            minLength: 0
            type: string
          minItems: 0
          type: array
      required:
      - converterType
      type: object
    removeFooterStructure:
      description: |
        This converter removes the footer from the table.
        The footer is defined as the last rows of the table which only contain a maximum of threshold invalid values.
        The invalid values are empty cells by default but can be extended using the blockList properties.

        The converterType MUST be REMOVE_FOOTER.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        threshold:
          description: The maximum number of invalid values in a row to be considered
            a footer row.
          minimum: 0
          type: integer
        blockList:
          description: The list of values to be considered invalid.
          items:
            minLength: 0
            type: string
          minItems: 0
          type: array
      required:
      - converterType
      type: object
...
    replaceEntriesStructure:
      description: |
        This converter replaces entries in the table with the specified replacement value.
        Entries to replace can be selected by specifying either a search string or a regex search string.
        If both search and regexSearch are provided, the search will be used.
        The replacement value will be used to replace the entries.
        This converter only affects the specified area of the table.

        The converterType MUST be REPLACE_ENTRIES.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
        search:
          description: The string to search for in the table.
          minLength: 0
          type: string
        regexSearch:
          description: The regex pattern to search for in the table.
          minLength: 1
          type: string
        replacement:
          description: The value to replace the found entries with.
          minLength: 0
          type: string
        columnIndex:
          description: The indices of the columns to replace
          items:
            minimum: 0
            type: integer
          minItems: 1
          type: array
        startRow:
          description: The start row index of the area to search in (inclusive).
          minimum: 0
          type: integer
        endRow:
          description: The end row index of the area to search in (exclusive).
          minimum: 0
          type: integer
      required:
      - converterType
      - replacement
      type: object
...
    tableStructureGenerationSettings:
      properties:
        removeHeader:
          $ref: '#/components/schemas/removeHeaderSettings'
        removeFooter:
          $ref: '#/components/schemas/removeFooterSettings'
        removeTrailingColumn:
          $ref: '#/components/schemas/removeTrailingColumnSettings'
        removeLeadingColumn:
          $ref: '#/components/schemas/removeLeadingColumnSettings'
        removeInvalidRows:
          $ref: '#/components/schemas/removeInvalidRowsSettings'
        removeKeywords:
          $ref: '#/components/schemas/removeKeywordsSettings'
        maxIterations:
          default: 5
          minimum: 1
          type: integer
      type: object
...
    report:
      discriminator:
        mapping:
          COLUMN_TYPE_MISMATCH: '#/components/schemas/columnTypeMismatchReport'
          GROUPED_HEADER: '#/components/schemas/groupedHeaderReport'
          EMPTY_HEADER: '#/components/schemas/emptyHeaderReport'
          EMPTY_ROW: '#/components/schemas/emptyRowReport'
          EMPTY_COLUMN: '#/components/schemas/emptyColumnReport'
          SAME_AS_HEADER: '#/components/schemas/sameAsHeaderReport'
          MISSING_ENTRY: '#/components/schemas/missingEntryReport'
          MERGEABLE_COLUMNS: '#/components/schemas/mergeableColumnsReport'
          SPLIT_ROW: '#/components/schemas/splitRowReport'
          SUM: '#/components/schemas/sumReport'
          TRANSPOSE_MATRIX: '#/components/schemas/transposeMatrixReport'
        propertyName: reportType
      example:
        reportType: EMPTY_COLUMN
        columnIndex: 42
      oneOf:
      - $ref: '#/components/schemas/columnTypeMismatchReport'
      - $ref: '#/components/schemas/groupedHeaderReport'
      - $ref: '#/components/schemas/emptyHeaderReport'
      - $ref: '#/components/schemas/emptyRowReport'
      - $ref: '#/components/schemas/emptyColumnReport'
      - $ref: '#/components/schemas/sameAsHeaderReport'
      - $ref: '#/components/schemas/missingEntryReport'
      - $ref: '#/components/schemas/mergeableColumnsReport'
      - $ref: '#/components/schemas/splitRowReport'
      - $ref: '#/components/schemas/sumReport'
      - $ref: '#/components/schemas/transposeMatrixReport'
      properties:
        reportType:
          $ref: '#/components/schemas/reportType'
      required:
      - reportType
      type: object
...
    tableStructureGenerationResponse:
      example:
        reports:
        - reportType: EMPTY_COLUMN
          columnIndex: 42
        - reportType: EMPTY_COLUMN
          columnIndex: 42
        tableStructure:
          id: 1
          name: Table structure 1
          structures:
          - name: Fill Header
            description: Fills empty cells in the header
            converterType: FILL_EMPTY_ROW
            rowIndex:
            - 0
            - 1
          - name: Resolve Grouped Header
            description: Resolves the grouped header
            converterType: REMOVE_GROUPED_HEADER
            rowIndex:
            - 0
            - 1
            columnIndex:
            - 0
            startRow: 3
            startColumn: 0
      properties:
        tableStructure:
          $ref: '#/components/schemas/tableStructure'
        reports:
          items:
            $ref: '#/components/schemas/report'
          type: array
      required:
      - reports
      - tableStructure
      type: object
...
    transposeMatrixStructure:
      description: |
        This converter transposes the given matrix.
      properties:
        converterType:
          $ref: '#/components/schemas/converterType'
      required:
      - converterType
      type: object
...
    generateTableStructure_request:
      properties:
        file:
          format: binary
          type: string
        settings:
          $ref: '#/components/schemas/tableStructureGenerationSettings'
      required:
      - file
      - settings
      type: object
    convertTable_request:
      properties:
        file:
          format: binary
          type: string
      required:
      - file
      type: object
    previewConvertTable_request:
      properties:
        file:
          format: binary
          type: string
        tableStructure:
          $ref: '#/components/schemas/tableStructure'
      required:
      - file
      - tableStructure
      type: object
  securitySchemes:
    oAuth2Auth:
      flows:
        authorizationCode:
          authorizationUrl: https://pg-doener-dev.virt.uni-oldenburg.de:8085/realms/th1/protocol/openid-connect/auth
          scopes:
            write:feedback: Scope to enable users to write feedback
            read:feedback: Scope to enable users to read feedback
            write:converter: Scope to enable users to write converter requests
            read:converter: Scope to enable users to read converter requests
            write:tablestructure: Scope to enable users to write table structures
            read:tablestructure: Scope to enable users to read table structures
          tokenUrl: https://pg-doener-dev.virt.uni-oldenburg.de:8085/realms/th1/protocol/openid-connect/token
      type: oauth2